restructure files for tests; test getLatestAvailableReports()
