+++ alerting/package.json
@@ -22,7 +22,7 @@
     "migration:revert": "yarn typeorm migration:revert",
     "typeorm-seeding": "ts-node $(yarn bin typeorm-seeding)",
     "alert-worker": "yarn ts-node src/alert-worker.ts",
-    "aa-storm-alert-worker": "yarn ts-node src/aa-storm-alert-worker.ts",
+    "aa-storm-alert-worker": "yarn ts-node src/aa-storm-alert-runner.ts",
     "docker-alert": "source ../api/set_envs.sh && docker compose run --entrypoint 'yarn alert-worker' alerting-node"
   },
   "dependencies": {
