+++ alerting/src/aa-storm-alert-runner.ts
@@ -1,0 +1,3 @@
+import { run } from './aa-storm-alert/worker';
+
+run();
