+++ alerting/src/aa-storm-alert/worker.ts
@@ -1,0 +1,45 @@
+import { createConnection, Repository } from 'typeorm';
+import { StormDataResponseBody, WindState } from '../types/rawStormDataTypes';
+import { LatestAAStormReports } from '../entities/latestAAStormReports.entity';
+import { StormAlertData } from '../types/email';
+import {
+  buildEmailPayloads,
+  filterAlreadyProcessedReports,
+  getLatestAvailableReports,
+} from './alert';
+
+// Replace with real function when available
+function sendStormAlertEmail(data: StormAlertData) {
+  //nothing yet
+}
+
+export async function run() {
+  // create a connection to the remote db
+  const connection = await createConnection();
+  const latestStormReportsRepository =
+    connection.getRepository(LatestAAStormReports);
+
+  const latestAvailableReports = await getLatestAvailableReports();
+
+  // filter reports which have been already processed
+
+  const filteredAvailableReports = await filterAlreadyProcessedReports(
+    latestAvailableReports,
+    latestStormReportsRepository,
+  );
+
+  // check whether an email should be sent
+  const emailPayloads = await buildEmailPayloads(filteredAvailableReports);
+
+  console.log('emailPayload', emailPayloads);
+
+  // create templates
+  sendStormAlertEmail(emailPayloads);
+  // send emails
+
+  // drop all latest storm reports stored to prevent accumulation of useless data in this table by time
+  await latestStormReportsRepository.clear();
+  await latestStormReportsRepository.save(
+    latestAvailableReports.map((report) => ({ reportIdentifier: report.path })),
+  );
+}
