+++ alerting/src/types/aa-storm-email.ts
@@ -1,0 +1,13 @@
+export interface EmailPayload {}
+
+export interface ShortReport {
+  ref_time: string;
+  state: string;
+  path: string;
+}
+
+export interface ShortReportsResponseBody {
+  [date: string]: {
+    [stormName: string]: ShortReport[];
+  };
+}
